requirements.txt

Django
django-cors-headers
djangorestframework
djangorestframework_simplejwt
filelock
gunicorn 
pytesseract
gunicorn
whitenoise
psycopg2-binary
dj-database-url
